Fix emergency behaviour by introducing double real_floor and removing previous_floor
The elevator would not go down when a order was placed below it after
an emergency. The variable real_floor is a double that is updated
to reflect where the elvator is between floors. I think real_floor
should replace current_floor, but that would require quite a bit of
change. The elevator is now working as intended.